Apply new app structure.
