+++ public/main.js
@@ -1,0 +1,9 @@
+import can from 'can';
+import $ from 'jquery';
+import stache from 'can/view/stache/stache';
+
+const template = stache('Hello {{message}}!');
+
+$(() => {
+  $('body').append(template({ message: 'Sergii' }));
+});
