--- public/main.stache
@@ -1,2 +1,0 @@
-<!-- main.stache -->
-<h1>{{message}}</h1>
