+++ public/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "public",
+  "version": "1.0.0",
+  "description": "",
+  "main": "main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Sergii Demianchuk",
+  "license": "ISC",
+  "dependencies": {
+    "can": "^2.3.23",
+    "jquery": "^2.2.3",
+    "steal": "^0.16.4"
+  }
+}
